login: move authentication and credential recovery into services/auth.py
Moves the token-minting helper, both login paths, password recovery,
verification-code sending, token validity checks and password reset out of
api/api_v1/endpoints/login.py and into a new services/auth.py, per
docs/proposals/2026-07-15-target-architecture.md step 4, item 2. The endpoints
now parse the request, resolve deps and call one service function.

Two of login.py's four endpoint-level db.commit() calls go with them:

  - reset_password's commit (was :359) was the last statement before the
    return, so relocating it is a no-op on the success path -- the request
    boundary in deps.get_db() commits instead.

  - _login_user's commit (was :84) is a genuine behavior change. It committed
    the `user.flags += " activated"` mutation *before* writing the audit log
    and minting the token. Deferring it to the single end-of-request commit is
    identical on success and differs only if a later step raises, in which case
    the flag no longer sticks. That is the atomicity hole of SS1.3 closing, as
    intended.

_verify_hcaptcha's body moves down too (it makes an outbound HTTP call); the
decision to require a captcha stays at level 1. Rate limits stay on the
endpoints. The pre-existing mis-ordering of @limiter.limit above @router.post
on recover_password -- which makes that limit inert -- is preserved and now
commented rather than silently fixed.

No HTTP contract change: paths, methods, response_models, function names
(operationId) and dependency shapes are all untouched.
