+++ chafan_core/app/services/__init__.py
@@ -7,6 +7,7 @@ from chafan_core.app.services import (
     article_columns,
     articles,
     audit,
+    auth,
     bot,
     channels,
     coin_deposits,
@@ -48,6 +49,7 @@ __all__ = [
     "article_columns",
     "articles",
     "audit",
+    "auth",
     "bot",
     "channels",
     "coin_deposits",
